+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Introduced `Framebuffer::builder` infrastructure
+  - Removed `Framebuffer::with_depth_map` constructor
 - Changed `sys::Gl::set_framebuffer_depth_texture` to
   `sys::Gl::set_framebuffer_texture` accepting
   `sys::FramebufferAttachment` enum